requirements.txt
streamlit
python-dotenv
pymysql
sentence-transformers
qdrant-client
groq
